refactor(team): extract OpenCode member inbox relay results, lease and delivery wake
TeamProvisioningOpenCodeMemberInboxRelay.ts mixed four unrelated concerns
in one 1155-line file. Move three of them out, in dependency order, without
changing any behaviour.

TeamProvisioningOpenCodeMemberInboxRelayResults.ts now owns the relay result
shape and its builders: OpenCodeMemberInboxRelayResult,
createOpenCodeMemberInboxRelayResult,
dedupeOpenCodeMemberInboxRelayDiagnostics and the seven
buildOpenCode...Result helpers.

TeamProvisioningOpenCodeMemberInboxRelayLease.ts now owns the in-flight
lease bookkeeping keyed off the inFlight map. Only
getActiveOpenCodeMemberInboxRelayWork and
registerOpenCodeMemberInboxRelayWork are exported; the lease record type,
the WeakMap store, the generation counter and the claim/release helpers stay
module-private exactly as they were.

TeamProvisioningOpenCodeMemberInboxDeliveryWake.ts now owns the watchdog
wake scheduling: OpenCodeMemberInboxDeliveryWakeInput,
OpenCodeMemberInboxDeliveryWakePorts and
scheduleOpenCodeMemberInboxDeliveryWakeWithPorts.

The moved code is byte-identical to its previous form; the only edits are
the `export` keyword added to the two lease functions the relay still calls.
The delivery-decision cluster (OpenCodeMemberInboxRelayDeliveryDecision,
handleOpenCodeInboxAttachmentFailure, projectOpenCodeInboxDeliveryFailure,
commitOpenCodeInboxRelayReadAfterDelivery) deliberately stays in the relay
file and keeps calling into the Results module, so the dependency direction
stays one-way and import/no-cycle is satisfied.

The relay file keeps its public surface: it re-exports the
OpenCodeMemberInboxRelayResult type for its nine importers and the
scheduleOpenCodeMemberInboxDeliveryWakeWithPorts function.
TeamProvisioningDiagnosticsPreflightCompatibilityFacade.ts imports the wake
module directly because it also needs the ports type. No test changed.

TeamProvisioningOpenCodeMemberInboxRelay.ts 1155 -> 871 lines; new modules
are 165, 103 and 38 lines.
